requirements.txt
#numpy==1.18.2
numpy==1.18.5
torch==1.4.0
tensorboardX==2.0
scikit_image==0.16.2
#scikit_image
matplotlib==2.2.3
tqdm==4.43.0
easydict==1.9
#Pillow==7.1.2
Pillow==8.3.2
opencv_python
pycocotools
